Refactor and enhance various components for improved structure and functionality
- Updated `page.tsx` to streamline data fetching for blog and manufacturer pages, integrating categories and authors.
- Refactored `BlogListClient` to include author filtering and improved category handling.
- Enhanced `ContactPage` to utilize a new `ContactClient` component for better organization and clarity.
- Simplified `ManufacturersPage` and `ManufacturerPage` to fetch and display manufacturer data more efficiently.
- Improved `ProjectsListClient` by adding a button for project discussions.
- Adjusted `TerritoryPage` to fetch territory information and render it using a dedicated client component.
- Updated Sanity client configuration to disable CDN for more accurate data during authoring.
- Enhanced schema definitions to include new types for manufacturers and territory information.
