+++ app/manufacturers/[slug]/ManufacturerDetailClient.tsx
@@ -1,0 +1,210 @@
+"use client"
+
+import { motion } from "framer-motion"
+import Image from "next/image"
+import Link from "next/link"
+import { ArrowLeft, ExternalLink, Phone } from "lucide-react"
+import { Badge } from "@/components/ui/badge"
+import { Button } from "@/components/ui/button"
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
+import { PageWrapper } from "@/components/page-wrapper"
+
+const fadeInUp = {
+  initial: { opacity: 0, y: 60 },
+  animate: { opacity: 1, y: 0 },
+  transition: { duration: 0.6 },
+}
+
+export type ManufacturerDetail = {
+  name: string
+  slug: string
+  logo?: string | null
+  category?: string | null
+  description?: string | null
+  keyProducts?: string[]
+  website?: string | null
+  specialty?: string | null
+  territoryNote?: string | null
+}
+
+type Props = {
+  manufacturer: ManufacturerDetail
+}
+
+export default function ManufacturerDetailClient({ manufacturer }: Props) {
+  const keyProducts = manufacturer.keyProducts || []
+
+  return (
+    <PageWrapper>
+      <section className="py-6 bg-slate-50">
+        <div className="container mx-auto px-4 lg:px-6">
+          <motion.div className="flex items-center space-x-2 text-sm" {...fadeInUp}>
+            <Link href="/manufacturers" className="flex items-center text-[#4986C8] hover:underline">
+              <ArrowLeft className="h-4 w-4 mr-1" />
+              Back to Manufacturers
+            </Link>
+          </motion.div>
+        </div>
+      </section>
+
+      <section className="py-10 md:py-14 bg-background">
+        <div className="container mx-auto px-4 lg:px-6">
+          <motion.div className="max-w-3xl mx-auto text-center" {...fadeInUp}>
+            {manufacturer.logo && (
+              <div className="mt-8 mb-6 flex justify-center">
+                <Image
+                  src={manufacturer.logo}
+                  alt={`${manufacturer.name} logo`}
+                  width={200}
+                  height={120}
+                  className="h-16 md:h-20 w-auto object-contain"
+                />
+              </div>
+            )}
+
+            <h1 className="text-4xl lg:text-5xl font-bold tracking-tight text-[#1C4E80] mb-2">
+              {manufacturer.name}
+            </h1>
+
+            {manufacturer.category && (
+              <div className="mb-4 text-center">
+                <Badge
+                  variant="outline"
+                  className="border-[#4986C8]/30 text-[#4986C8] bg-blue-50 px-3 py-1 inline-block"
+                >
+                  {manufacturer.category}
+                </Badge>
+              </div>
+            )}
+
+            {manufacturer.description && (
+              <p className="text-lg text-gray-700 leading-relaxed mb-6 max-w-2xl mx-auto">
+                {manufacturer.description}
+              </p>
+            )}
+
+            {manufacturer.specialty && (
+              <p className="text-base text-muted-foreground max-w-xl mx-auto mb-4">
+                Specialty: {manufacturer.specialty}
+              </p>
+            )}
+
+            {manufacturer.territoryNote && (
+              <div className="bg-[#4986C8]/10 border border-[#4986C8]/30 rounded-lg p-4 mx-auto mb-6 max-w-2xl">
+                <p className="text-[#1C4E80] font-semibold">Territory Note</p>
+                <p className="text-[#123D6A] text-sm">{manufacturer.territoryNote}</p>
+              </div>
+            )}
+
+            <div className="flex flex-col md:flex-row gap-4 justify-center">
+              <Link href="/contact">
+                <Button size="lg" className="bg-[#4986C8] hover:bg-[#4986C8]/90 px-8 py-3">
+                  <Phone className="mr-2 h-4 w-4" />
+                  Contact Rep
+                </Button>
+              </Link>
+              {manufacturer.website && (
+                <Button
+                  size="lg"
+                  variant="outline"
+                  className="bg-transparent border-2 border-[#4986C8] text-[#4986C8] hover:bg-[#4986C8] hover:text-white px-8 py-3 font-medium"
+                  asChild
+                >
+                  <a href={manufacturer.website} target="_blank" rel="noopener noreferrer">
+                    <ExternalLink className="mr-2 h-4 w-4" />
+                    Visit Website
+                  </a>
+                </Button>
+              )}
+            </div>
+          </motion.div>
+        </div>
+      </section>
+
+      <section className="py-10 md:py-14 bg-slate-50">
+        <div className="container mx-auto px-4 lg:px-6">
+          <div className="max-w-6xl mx-auto">
+            <motion.div className="text-center space-y-4 mb-12" {...fadeInUp}>
+              <h2 className="text-3xl lg:text-4xl font-bold text-[#1C4E80]">Key Products</h2>
+              <p className="text-xl text-muted-foreground max-w-3xl mx-auto">
+                Product highlights from {manufacturer.name}
+              </p>
+            </motion.div>
+
+            {keyProducts.length === 0 ? (
+              <p className="text-center text-muted-foreground">No products listed yet.</p>
+            ) : (
+              <div className="space-y-4">
+                {keyProducts.map((product, index) => {
+                  const isHeader = product.includes("<b>")
+                  if (isHeader) {
+                    return (
+                      <motion.div
+                        key={index}
+                        initial={{ opacity: 0, y: 30 }}
+                        whileInView={{ opacity: 1, y: 0 }}
+                        transition={{ duration: 0.6, delay: index * 0.05 }}
+                        viewport={{ once: true }}
+                      >
+                        <h3
+                          className="text-xl font-bold text-[#1C4E80] mt-8 mb-4 first:mt-0"
+                          dangerouslySetInnerHTML={{ __html: product }}
+                        />
+                      </motion.div>
+                    )
+                  }
+
+                  return (
+                    <motion.div
+                      key={index}
+                      initial={{ opacity: 0, y: 30 }}
+                      whileInView={{ opacity: 1, y: 0 }}
+                      transition={{ duration: 0.6, delay: index * 0.05 }}
+                      viewport={{ once: true }}
+                    >
+                      <Card className="hover:shadow-lg transition-shadow">
+                        <CardContent className="p-6">
+                          <p className="text-[#123D6A] font-medium leading-relaxed">{product}</p>
+                        </CardContent>
+                      </Card>
+                    </motion.div>
+                  )
+                })}
+              </div>
+            )}
+          </div>
+        </div>
+      </section>
+
+      <section className="py-16 bg-background">
+        <div className="container mx-auto px-4 lg:px-6">
+          <Card className="p-6 md:p-10 bg-[#123D6A] text-white">
+            <CardHeader className="px-0">
+              <CardTitle className="text-2xl md:text-3xl">
+                Need help with {manufacturer.name} equipment?
+              </CardTitle>
+            </CardHeader>
+            <CardContent className="px-0 space-y-4">
+              <p className="opacity-90">
+                Our team can guide you to the right products and specifications for your project.
+              </p>
+              <div className="flex flex-col sm:flex-row gap-3">
+                    <Link href="/contact">
+                      <Button size="lg" className="bg-[#4986C8] hover:bg-[#4986C8]/90">Contact Your Rep</Button>
+                    </Link>
+                {manufacturer.website && (
+                  <Button size="lg" variant="outline" className="border-white text-white hover:bg-white hover:text-[#123D6A] bg-transparent" asChild>
+                    <a href={manufacturer.website} target="_blank" rel="noopener noreferrer">
+                      Visit Website
+                    </a>
+                  </Button>
+                )}
+              </div>
+            </CardContent>
+          </Card>
+        </div>
+      </section>
+    </PageWrapper>
+  )
+}
+
