+++ lib/sanity.queries.ts
@@ -55,3 +55,84 @@ export const projectsListQuery = groq`*[_type == "project"]|order(coalesce(publi
   "imageUrl": mainImage.asset->url,
   categories[]->{ title }
 }`
+
+export const allManufacturerSlugsQuery = groq`*[_type == "manufacturer" && defined(slug.current)]{ "slug": slug.current }`
+
+export const manufacturersListQuery = groq`*[_type == "manufacturer"]|order(coalesce(order, 9999) asc, name asc){
+  _id,
+  name,
+  "slug": slug.current,
+  category,
+  description,
+  specialty,
+  keyProducts,
+  website,
+  territoryNote,
+  featured,
+  order,
+  logo,
+  "logoUrl": logo.asset->url
+}`
+
+export const manufacturerBySlugQuery = groq`*[_type == "manufacturer" && slug.current == $slug][0]{
+  _id,
+  _updatedAt,
+  name,
+  "slug": slug.current,
+  category,
+  description,
+  specialty,
+  keyProducts,
+  website,
+  territoryNote,
+  featured,
+  order,
+  logo,
+  "logoUrl": logo.asset->url
+}`
+
+export const siteSettingsQuery = groq`*[_type == "siteSettings"][0]{
+  heroBadge,
+  heroTitle,
+  heroDescription,
+  heroImage,
+  primaryCta,
+  secondaryCta,
+  stats,
+  manufacturerStrip[]->{
+    name,
+    "slug": slug.current,
+    category,
+    logo,
+    "logoUrl": logo.asset->url
+  },
+  highlights[]{
+    title,
+    description,
+    category,
+    states,
+    image
+  }
+}`
+
+export const representativesQuery = groq`*[_type == "representative"]|order(coalesce(order, 9999) asc, name asc){
+  name,
+  role,
+  phone,
+  email,
+  states,
+  regions
+}`
+
+export const territoryInfoQuery = groq`*[_type == "territoryInfo"][0]{
+  heroTitle,
+  heroSubtitle,
+  primaryCta,
+  secondaryCta,
+  coverageBlurb,
+  businessHours
+}`
+
+export const categoriesListQuery = groq`*[_type == "category"]{ title }`
+
+export const authorsListQuery = groq`*[_type == "author"]{ name }`
