+++ sanity/schemaTypes/manufacturer.ts
@@ -1,0 +1,79 @@
+import { defineField, defineType } from 'sanity'
+
+export default defineType({
+  name: 'manufacturer',
+  title: 'Manufacturer',
+  type: 'document',
+  fields: [
+    defineField({
+      name: 'name',
+      title: 'Name',
+      type: 'string',
+      validation: (rule) => rule.required(),
+    }),
+    defineField({
+      name: 'slug',
+      title: 'Slug',
+      type: 'slug',
+      options: {
+        source: 'name',
+        maxLength: 96,
+      },
+      validation: (rule) => rule.required(),
+    }),
+    defineField({
+      name: 'logo',
+      title: 'Logo',
+      type: 'image',
+      options: { hotspot: true },
+    }),
+    defineField({
+      name: 'category',
+      title: 'Category',
+      type: 'string',
+      validation: (rule) => rule.required(),
+    }),
+    defineField({
+      name: 'description',
+      title: 'Description',
+      type: 'text',
+    }),
+    defineField({
+      name: 'specialty',
+      title: 'Specialty',
+      type: 'string',
+      description: 'Short highlight such as key segment or product specialty',
+    }),
+    defineField({
+      name: 'keyProducts',
+      title: 'Key Products',
+      type: 'array',
+      of: [{ type: 'string' }],
+    }),
+    defineField({
+      name: 'website',
+      title: 'Website',
+      type: 'url',
+    }),
+    defineField({
+      name: 'territoryNote',
+      title: 'Territory Note',
+      type: 'text',
+      rows: 2,
+      description: 'Optional note about territory restrictions',
+    }),
+    defineField({
+      name: 'featured',
+      title: 'Featured',
+      type: 'boolean',
+      initialValue: false,
+    }),
+    defineField({
+      name: 'order',
+      title: 'Order',
+      type: 'number',
+      description: 'Lower numbers appear first',
+    }),
+  ],
+})
+
