+++ sanity/schemaTypes/siteSettings.ts
@@ -1,0 +1,87 @@
+import { defineField, defineType } from 'sanity'
+
+export default defineType({
+  name: 'siteSettings',
+  title: 'Site Settings',
+  type: 'document',
+  fields: [
+    defineField({
+      name: 'heroBadge',
+      title: 'Hero Badge',
+      type: 'string',
+    }),
+    defineField({
+      name: 'heroTitle',
+      title: 'Hero Title',
+      type: 'string',
+    }),
+    defineField({
+      name: 'heroDescription',
+      title: 'Hero Description',
+      type: 'text',
+    }),
+    defineField({
+      name: 'heroImage',
+      title: 'Hero Image',
+      type: 'image',
+      options: { hotspot: true },
+    }),
+    defineField({
+      name: 'primaryCta',
+      title: 'Primary CTA',
+      type: 'object',
+      fields: [
+        defineField({ name: 'label', title: 'Label', type: 'string' }),
+        defineField({ name: 'href', title: 'Href', type: 'string' }),
+      ],
+    }),
+    defineField({
+      name: 'secondaryCta',
+      title: 'Secondary CTA',
+      type: 'object',
+      fields: [
+        defineField({ name: 'label', title: 'Label', type: 'string' }),
+        defineField({ name: 'href', title: 'Href', type: 'string' }),
+      ],
+    }),
+    defineField({
+      name: 'stats',
+      title: 'Stats',
+      type: 'array',
+      of: [
+        {
+          type: 'object',
+          fields: [
+            defineField({ name: 'value', title: 'Value', type: 'string' }),
+            defineField({ name: 'label', title: 'Label', type: 'string' }),
+            defineField({ name: 'detail', title: 'Detail', type: 'string' }),
+          ],
+        },
+      ],
+    }),
+    defineField({
+      name: 'manufacturerStrip',
+      title: 'Manufacturer Strip',
+      type: 'array',
+      of: [{ type: 'reference', to: [{ type: 'manufacturer' }] }],
+    }),
+    defineField({
+      name: 'highlights',
+      title: 'Project Highlights',
+      type: 'array',
+      of: [
+        {
+          type: 'object',
+          fields: [
+            defineField({ name: 'title', title: 'Title', type: 'string' }),
+            defineField({ name: 'description', title: 'Description', type: 'text' }),
+            defineField({ name: 'category', title: 'Category', type: 'string' }),
+            defineField({ name: 'states', title: 'States', type: 'array', of: [{ type: 'string' }] }),
+            defineField({ name: 'image', title: 'Image', type: 'image', options: { hotspot: true } }),
+          ],
+        },
+      ],
+    }),
+  ],
+})
+
