+++ sanity/schemaTypes/territoryInfo.ts
@@ -1,0 +1,49 @@
+import { defineField, defineType } from 'sanity'
+
+export default defineType({
+  name: 'territoryInfo',
+  title: 'Territory Info',
+  type: 'document',
+  fields: [
+    defineField({ name: 'heroTitle', title: 'Hero Title', type: 'string' }),
+    defineField({ name: 'heroSubtitle', title: 'Hero Subtitle', type: 'text' }),
+    defineField({
+      name: 'primaryCta',
+      title: 'Primary CTA',
+      type: 'object',
+      fields: [
+        defineField({ name: 'label', title: 'Label', type: 'string' }),
+        defineField({ name: 'href', title: 'Href', type: 'string' }),
+      ],
+    }),
+    defineField({
+      name: 'secondaryCta',
+      title: 'Secondary CTA',
+      type: 'object',
+      fields: [
+        defineField({ name: 'label', title: 'Label', type: 'string' }),
+        defineField({ name: 'href', title: 'Href', type: 'string' }),
+      ],
+    }),
+    defineField({
+      name: 'coverageBlurb',
+      title: 'Coverage Blurb',
+      type: 'text',
+    }),
+    defineField({
+      name: 'businessHours',
+      title: 'Business Hours',
+      type: 'array',
+      of: [
+        {
+          type: 'object',
+          fields: [
+            defineField({ name: 'label', title: 'Label', type: 'string' }),
+            defineField({ name: 'value', title: 'Value', type: 'string' }),
+          ],
+        },
+      ],
+    }),
+  ],
+})
+
